+++ nodejs/awsx/aws_test.go
@@ -102,6 +102,11 @@ func Test_Examples(t *testing.T) {
 			},
 		}),
 		testBase.With(integration.ProgramTestOptions{
+			Dir:                    path.Join(cwd, "../examples/nlb/fargateShortInlineListener"),
+			StackName:              addRandomSuffix("fargate"),
+			ExtraRuntimeValidation: containersRuntimeValidator(envRegion, true /*isFargate*/, true /*short*/),
+		}),
+		testBase.With(integration.ProgramTestOptions{
 			Dir: path.Join(cwd, "../examples/api"),
 			Dependencies: []string{
 				"@pulumi/awsx",
@@ -218,7 +223,15 @@ func Test_Examples(t *testing.T) {
 			},
 		}),
 		testBase.With(integration.ProgramTestOptions{
+			Dir:       path.Join(cwd, "../examples/alb/fargateInlineListener"),
+			StackName: addRandomSuffix("fargate"),
+		}),
+		testBase.With(integration.ProgramTestOptions{
 			Dir:       path.Join(cwd, "../examples/alb/ec2"),
+			StackName: addRandomSuffix("ec2"),
+		}),
+		testBase.With(integration.ProgramTestOptions{
+			Dir:       path.Join(cwd, "../examples/alb/ec2InlineListener"),
 			StackName: addRandomSuffix("ec2"),
 		}),
 		testBase.With(integration.ProgramTestOptions{
@@ -231,6 +244,19 @@ func Test_Examples(t *testing.T) {
 		}),
 		testBase.With(integration.ProgramTestOptions{
 			Dir:       path.Join(cwd, "../examples/nlb/fargate"),
+			StackName: addRandomSuffix("fargate"),
+			Config: map[string]string{
+				"aws:region":               "INVALID_REGION",
+				"aws:envRegion":            envRegion,
+				"containers:redisPassword": "SECRETPASSWORD",
+			},
+			PreviewCommandlineFlags: []string{
+				"--diff",
+			},
+			ExtraRuntimeValidation: containersRuntimeValidator(envRegion, true /*isFargate*/, false /*short*/),
+		}),
+		testBase.With(integration.ProgramTestOptions{
+			Dir:       path.Join(cwd, "../examples/nlb/fargateInlineListener"),
 			StackName: addRandomSuffix("fargate"),
 			Config: map[string]string{
 				"aws:region":               "INVALID_REGION",